test: prune redundant csv-import e2e tests and merge small suites
